requirements.txt
requests~=2.32.3
urllib3~=2.3.0
tqdm~=4.67.1
validators~=0.34.0
cryptography~=44.0.0
python-dotenv~=1.0.1